+++ components/announcement/list/Banner.tsx
